refactor: move utilities to shared module
