+++ CONTEXT09052025.md
@@ -18,8 +18,8 @@ Kepner-Tregoe’s Certs & Badges System (CBS) manages workshops (“Sessions”)
 0.6 Update this file when you finish work (flip items to **[DONE]**).  
 0.7 **Static assets:** Do **not** add/replace images under `app/assets/badges/`. Use what’s already there. Badges are served from `/srv/badges`. No binary churn in PRs.  
 0.8 **Scope guard:** Don’t invent pages or routes. Follow sections below.
-0.9 Utility modules live under `app/utils/`; import them as `from app.utils`.
-    - Imports should use `app.utils.*` (top-level `utils/` was removed).
+0.9 Utility modules live under `app/shared/`; import them as `from app.shared`.
+    - Imports should use `app.shared.*` (top-level `utils/` was removed).
 
 ---
