+++ app/shared/acl.py
@@ -5,7 +5,7 @@ from zoneinfo import ZoneInfo
 from typing import Any
 
 from ..models import Session, User, ParticipantAccount
-from ..constants import (
+from .constants import (
     SYS_ADMIN,
     ADMIN,
     CONTRACTOR,
