+++ app/shared/auth_bridge.py
