+++ app/shared/badges.py
