+++ app/shared/constants.py
