+++ app/shared/strings.py
