+++ app/shared/views.py
