+++ tests/test_time_format.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from app.utils.time import fmt_dt
+from app.shared.time import fmt_dt
 
 def test_fmt_dt_no_seconds():
     dt = datetime(2024, 1, 2, 3, 4, 5)
